% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutPoly.R
\name{cutPoly}
\alias{cutPoly}
\title{cutPoly}
\usage{
cutPoly(sPoly, xPoly)
}
\arguments{
\item{sPoly}{\code{sf} object, the input polygon within which line transects
will be generated. The coordinate system
must use meters as the linear unit (e.g. EPSG=26913)}

\item{xPoly}{\code{sf} object, optional input polygon that defines an
exclusion area where line transects will not be generated.}
}
\value{
A polygon with cutouts
}
\description{
This function cuts out holes in a given polygon by transforming the crs' to
match and then by differencing the 2 polygons.
}
\examples{
\dontrun{
library(sf)
library(dplyr)

# get hunt area polygons
poly <- huntAreas \%>\%
  filter(HERDNAME \%in\% c("Badwater", "North Natrona"))

# map
mapLT(poly)

# get cutout
x <- st_read("example_cutout.kml")

# map poly and cutouts
mapLT(poly, x)

# get poly with cutouts
poly <- cutPoly(poly, x)

# calculate recommended line length
recLength <- calcLineLength(
  poly,
  N = 20000
)

# makeLines
lines <- makeLines(
  poly,
  recLength
)

# map
mapLT(poly, x, lines$lines)
}

}
\author{
Garrett Catlin
}
