% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistSampModel.R
\name{fitDistSampModel}
\alias{fitDistSampModel}
\title{Fit distance-sampling model to pronghorn line-transect data}
\usage{
fitDistSampModel(
  ddf,
  sdf,
  keyFun = "halfnorm",
  wHi = 200,
  sidesSurveyed = 1,
  areaMi2,
  bootIterations
)
}
\arguments{
\item{ddf}{Detection data.  See \code{Rdistance::dfuncEstim} and
\code{Rdistance::abundEstim}.}

\item{sdf}{Site data.  See \code{Rdistance::abundEstim}.}

\item{keyFun}{Shape of detection curve.  See \code{Rdistance::dfuncEstim}.}

\item{wHi}{Right truncation distance.  See \code{Rdistance::dfuncEstim}.
Default is 200 m, given the pronghorn survey strip is 200 m wide.}

\item{sidesSurveyed}{Sides of the aircraft surveyed.  Default is 1, given the
current protocol specifies 1 observer looking out 1 side of the aircraft.
Rdistance assumes 2 sides of the line are surveyed, so this allows the survey
strip area to be adjusted accordingly.}

\item{areaMi2}{Size of the area (in square miles) that the abundance estimate
is extrapolated to.  This is often the total area of the herd unit, but areas
not occupiable by pronghorn (e.g., forested areas) should be excluded.}

\item{bootIterations}{Number of bootstrap iterations to use to calculate
confidence intervals.  See \code{Rdistance::abundEstim}.}
}
\value{
An 'abundance estimate' object.  Same as returned by
\code{Rdistance::abundEstim}.
}
\description{
Estimate abundance of pronghorn using distance sampling.
Distance sampling models are fit using Rdistance, and this code
is compatible with Rdistance version 2.2.0.
This is a thin wrapper of \code{Rdistance::dfuncEstim} and
\code{Rdistance::abundEstim}.
}
\examples{
\dontrun{
# Prep data
dataPath <- "C:/Users/jadcarlisle/Desktop/demo"
x <- prepDataForAnalysis(inputFile = file.path(dataPath,
                                              "Data_PronghornLT_Rattlesnake_2022.xlsx"),
                       inputSheet = 1,
                       shpCreate = FALSE,
                        mapCreate= FALSE)

# Fit distance-sampling model to estimate abundance
# Key input data are the x$ddf and x$sdf data.frames from prepDataForAnalysis
fit <- fitDistSampModel(ddf = x$ddf,
                       sdf = x$sdf,
                       keyFun = "hazrate",
                       sidesSurveyed = 1,
                       areaMi2 = 884,
                       bootIterations = 50)

# Print results
fit
}
}
\author{
Jason Carlisle
}
