% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herdUnits.R
\docType{data}
\name{herdUnits}
\alias{herdUnits}
\title{Herd Units in Wyoming}
\format{
An sf data frame with herd units in Wyoming
}
\usage{
herdUnits
}
\description{
Herd Units in Wyoming
}
\keyword{datasets}
