% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huntAreas.R
\docType{data}
\name{huntAreas}
\alias{huntAreas}
\title{Pronghorn/antelope Hunt Areas in Wyoming}
\format{
An sf data frame of 102 pronghorn/antelope hunt areas in Wyoming,
current as of May 2023.
}
\usage{
huntAreas
}
\description{
Pronghorn/antelope Hunt Areas in Wyoming
}
\keyword{datasets}
