% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLines.R
\name{makeLines}
\alias{makeLines}
\title{makeLines}
\usage{
makeLines(
  sPoly,
  xPoly = NULL,
  totalLengthKm,
  angle = 90,
  minLengthKm = 2,
  offset = 0,
  minSpace = NULL,
  maxSpace = NULL,
  optimTol = 0.01,
  outFile = NULL
)
}
\arguments{
\item{sPoly}{The input polygon within which line transects will be generated.
This must be a sf object and the coordinate system must us meters as the
linear unit (e.g. EPSG=26913)}

\item{xPoly}{Optional input polygon that defines an exclusion area where
line transects will not be generated. This must be a sf object.}

\item{totalLengthKm}{Total length of transect to generate (in kilometers)}

\item{angle}{The angle defining the orientation of the transect. 
\code{angle = 90} corresponds to transects oriented in a North-South 
direction (default)}

\item{minLengthKm}{Minimum length required for an indiviudal transect line
(in kilometers)}

\item{offset}{an offset to use when placing transects. By default the central
transect is aligned with the center}

\item{minSpace}{Optional minimal line spacing (in kilometers) to consider when
optimizing transect placement (default is NULL).}

\item{maxSpace}{Optional maximum line spacing (in kilometers) to consider when
optimizing transect placement (default is NULL).}

\item{optimTol}{Optional tolerance value indicating the maximum allowable
deviation from \code{totalLengthKm} expressed as a proportion. (default is 0.01).
When the function fails to find a solution within the specified tolerance an
error is returned.}

\item{outFile}{Optional output filename for a saving a spatial layer containing
the line transects. Typically, this will be a shapefile (e.g. "C:/Temp/myLines.shp")
The file extenstion can be any accepted by \code{sf::st_write()}}
}
\value{

}
\description{
This function optimizes the placement and spacing of 
distance sampling transects for a given polygon area and desired total 
transect length.
}
\examples{
\dontrun{
#Read in herd unit polygon
hu <- sf::st_read("W:/My Drive/Projects/Pronghorn_ltds/herd_units/Rattlesnake_HU.shp")

# Make lines
testLines <- makeLines(sPoly = hu,
     xPoly = NULL,
     totalLengthKm = 1000,
     angle = 90,
     outFile = NULL,
     minSpace = NULL,
     maxSpace = NULL)
     
# View output summary
print(testLines$summary)

# Plot output
plot(hu["geometry"])
plot(testLines$lines['lineID'], add = T, col = 'red')
}
}
\author{
Thomas J. Prebyl and Jason D. Carlisle, \email{jason.carlisle@wyo.gov }
}
