% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapLT.R
\name{mapLT}
\alias{mapLT}
\title{mapLT}
\usage{
mapLT(herdPoly, cutoutPoly = NULL, lines = NULL)
}
\arguments{
\item{herdPoly}{sf object, herd unit and/or hunt area boundary (sample frame).}

\item{cutoutPoly}{optional sf object, area excluded from sample frame.}

\item{lines}{optional sf object generated from \code{\link{makeLines}}, transect lines.}
}
\description{
Maps hunt area and/or herd unit polygons, cutouts (if applicable), and survey lines (if applicable).
}
\examples{
\dontrun{
library(sf)
library(dplyr)

# get hunt area polygons
poly <- huntAreas \%>\%
  filter(HERDNAME \%in\% c("Badwater", "North Natrona"))

# map
mapLT(poly)

# get cutout
x <- st_read("example_cutout.kml")

# map poly and cutouts
mapLT(poly, x)

# get poly with cutouts
poly <- cutPoly(poly, x)

# calculate recommended line length
recLength <- calcLineLength(
  poly,
  N = 20000
)

# makeLines
lines <- makeLines(
  poly,
  recLength
)

# map
mapLT(poly, x, lines$lines)
}

}
\author{
Garrett Catlin
}
