% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapResults.R
\name{mapResults}
\alias{mapResults}
\title{mapResults}
\usage{
mapResults(preppedData)
}
\arguments{
\item{preppedData}{List object returned from \code{\link{prepDataForAnalysis}}.}
}
\value{
Returns a leaflet map.
}
\description{
Maps the results of a survey.
}
\examples{
\dontrun{

# Prep data
dat <- prepDataForAnalysis("example.xlsx")

# Map
mapResults(dat)
}

}
\author{
Garrett Catlin
}
