% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDataForAnalysis.R
\name{prepDataForAnalysis}
\alias{prepDataForAnalysis}
\title{prepDataForAnalysis}
\usage{
prepDataForAnalysis(inputFile = NULL, inputSheet = NULL, outputFile = NULL)
}
\arguments{
\item{inputFile}{The data output from CyberTracker to format for input to
Program DISTANCE.  Excel files with extension \code{.xls} or \code{.xlsx} are
allowed. If \code{inputFile = NULL} (the default), the user is
prompted to select the input file interactively. Alternatively, the user can
specify the path and name of the Excel file containing the input data.
Note the direction of slashes in the path must follow R expectations
(use \code{/} instead of the Windows default of \code{\\}),
that the file name and extension must be included,
and that the path must start and end with quotes.
Example:  \code{"C:/Users/jcarlisle/Desktop/myInputFile.xlsx"}.
This function expects the following columns (named as shown) in the
\code{inputFile}.  The order of the columns does not matter (except in the
case of the "Cluster size" columns, see below), and
any additional columns are ignored:
\describe{
 \item{Transect number}{Unique name of each transect.}
 \item{Flight control}{Indicates transect starts/stops from detections.}
 \item{Latitude}{Y coordinate from GPS.}
 \item{Longitude}{X coordinate from GPS.}
 \item{Range}{Height of airplane (above ground level [AGL]) in feet.}
 \item{Cluster size}{For some reason, the CyberTracker output contains two
 columns named "Cluster size".  This function expects the first one to be
 the number of pronghorn in the detected group.}
 \item{Cluster size}{For some reason, the CyberTracker output contains two
 columns named "Cluster size".  This function expects the second one to be
 the distance band the pronghorn were detected in.}
}}

\item{inputSheet}{The sheet (or tab) within the above-specified Excel file
that should be read in. If \code{inputSheet = NULL} (the default), the names
of the sheets in the \code{inputFile} are printed to the console, and the user
is prompted to select the sheet name interactively.  Alternatively,
specify as the name of the sheet (e.g., \code{"Sheet1"} or \code{"inputData"})
or as the numbered position of the sheet within the file (e.g., \code{1}).}

\item{outputFile}{The path and name of the tab-delimited text file to write
out containing the output data ready for import into Program DISTANCE.
If \code{outputFile = NULL} (the default), the file will be written to the
same directory as the \code{inputFile}, with \code{"PreppedData_"} prepended
to the file name of the \code{inputFile}.  E.g, if
\code{inputFile} is \code{"C:/Users/jcarlisle/Desktop/myInputFile.xlsx"} and
\code{outputFile = NULL}, the output will be written to
\code{"C:/Users/jcarlisle/Desktop/PreppedData_myInputFile.txt"}.
Alternatively, the user can specify the path and name of the text file.
Note the direction of slashes in the path must follow R expectations
(use \code{/} instead of the Windows default of \code{\\}),
that the file name and extension must be included,
and that the path must start and end with quotes.
#' Example:  \code{"C:/Users/jcarlisle/Desktop/myOutputFile.txt"}.
The column names in the output file are standardized to use names similar to
Program DISTANCE (making import less error-prone) and to clarify units:
\describe{
\item{lineLabel}{Unique name of each transect.}
\item{lineLengthKm}{Length of each transect in kilometers.}
\item{coordsX}{X coordinate from GPS.}
\item{coordsY}{Y coordinate from GPS.}
\item{flightHeightFt}{Height of airplane (AGL) in feet.}
\item{clusterSize}{Number of pronghorn in detected group.}
\item{distBand}{Distance band the pronghorn were detected in.}
\item{adjustedDistM}{Estimated distance from transect to the detected
pronghorn group in meters. Based on the midpoint of the distance band
containing the detected group, adjusted for the flight height at the time of
detection.}
}}
}
\value{
Writes a tab-delimited text file to \code{outputFile}.  This file is
ready for import into Program DISTANCE.  Also prints the
following summaries in the R console:
\itemize{
\item Number of transects.
\item Total transect length surveyed (km).
\item Number of groups (clusters) detected.
\item Assuming right-truncation at largest adjusted distance interval cutpoint, number of groups (clusters) in analysis.
\item Number of individuals detected.
\item Assuming right-truncation at largest adjusted distance interval cutpoint, number of individuals in analysis.
\item Number of groups (clusters) that had missing flight heights
\item Mean flight height (ft) at detections - before imputing missing values
\item Mean flight height (ft) at detections - after imputing missing values
\item Nominal distance interval cutpoints (m)
\item Adjusted distance interval cutpoints (m)
}
}
\description{
This function reads in raw data exported from CyberTracker
software and prepares it for input to Program DISTANCE for a
distance-sampling analysis.
The prepped data is written to file, and some summaries of the data are
printed to the R console.
}
\examples{
\dontrun{
# Read in Sheet1 of myInputFile.xlsx, and write out myOutputFile.txt
prepDataForAnalysis(inputFile = "C:/Users/myUserName/Desktop/myInputFile.xlsx",
                    inputSheet = "Sheet1",
                    outputFile = "C:/Users/myUserName/Desktop/myOutputFile.txt")
}
}
\references{
Guenzel, R.J. 2007. Procedures for Estimating Pronghorn Abundance
in Wyoming Using Aerial Line Transect Sampling. Wyoming Game and Fish Department.
Cheyenne, WY, USA.
}
\author{
Jason D. Carlisle, \email{jcarlisle@west-inc.com}
}
