% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herdUnits.R
\docType{data}
\name{herdUnits}
\alias{herdUnits}
\title{Pronghorn/antelope Herd Units in Wyoming}
\format{
An sf data frame of 40 pronghorn/antelope herd units in Wyoming,
current as of fall 2022.
}
\usage{
herdUnits
}
\description{
Pronghorn/antelope Herd Units in Wyoming
}
\keyword{datasets}
