% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huntAreas.R
\docType{data}
\name{huntAreas}
\alias{huntAreas}
\title{Hunt Areas in Wyoming}
\format{
An sf data frame with hunt areas in Wyoming
}
\usage{
huntAreas
}
\description{
Hunt Areas in Wyoming
}
\keyword{datasets}
