% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLines.R
\name{makeLines}
\alias{makeLines}
\title{makeLines}
\usage{
makeLines(
  sPoly,
  totalLengthKm,
  angle = 0,
  minLengthKm = 2,
  offset = 0,
  considerHuntArea = TRUE,
  huntAreaBuffer = 1000,
  breakLongLines = TRUE,
  breakLongLineLength = 40,
  minSpace = NULL,
  maxSpace = NULL,
  optimTol = 0.01,
  optimQuick = FALSE
)
}
\arguments{
\item{sPoly}{\code{sf} object, the input polygon within which line transects
will be generated. The coordinate system
must use meters as the linear unit (e.g. EPSG=26913)}

\item{totalLengthKm}{Total length of transects to generate (in kilometers)}

\item{angle}{Orientation of the transects. \code{angle = 0} produces
transects oriented North-South (the default), \code{angle = 90} produces
East-West transects.}

\item{minLengthKm}{Minimum length (km) required for an individual transect
line to be surveyed.}

\item{offset}{An offset to use when placing transects. By default the central
transect is aligned with the center of the polygon.}

\item{considerHuntArea}{Breaks transects as they cross over hunt areas
(defaults to TRUE).}

\item{huntAreaBuffer}{Buffer by which to break transects when considerHuntArea
is TRUE (defaults to 1000 km).}

\item{breakLongLines}{Splits longer transects into more digestible pieces
(defaults to TRUE).}

\item{breakLongLineLength}{Limit for maximum line length (in km) if
breakLongLines is TRUE (defaults to 40 km).}

\item{minSpace}{Optional minimal line spacing (in km) to consider when
optimizing transect placement (default is NULL).}

\item{maxSpace}{Optional maximum line spacing (in km) to consider when
optimizing transect placement (default is NULL).}

\item{optimTol}{Optional tolerance value indicating the maximum allowable
deviation from \code{totalLengthKm} expressed as a proportion (default is 0.01).
When the function fails to find a solution within the specified tolerance an
error is returned.}

\item{optimQuick}{if TRUE, the initial optimization is used. Helpful for when
calculations need to fail quickly.}
}
\value{
Returns a list object containing the transect lines and
a summary dataframe:
\describe{
  \item{lines}{\code{sf} object, simple feature collection of transect lines}
  \item{summary}{\code{data.frame} summary of transect lines}
}
}
\description{
This function optimizes the placement and spacing of
line transects for a given polygon area and desired total
transect length.
}
\examples{
\dontrun{
#Read in herd unit polygon
hu <- sf::st_read("W:/My Drive/Projects/Pronghorn_ltds/herd_units/Rattlesnake_HU.shp")

# Make lines
testLines <- makeLines(sPoly = hu,
     totalLengthKm = 1000,
     angle = 0,
     minSpace = NULL,
     maxSpace = NULL)

# View output summary
print(testLines$summary)

# Plot output
plot(hu["geometry"])
plot(testLines$lines['lineID'], add = T, col = 'red')
}
}
\author{
Tom Prebyl, Jason Carlisle, and Garrett Catlin
}
