% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLines.R
\name{makeLines}
\alias{makeLines}
\title{makeLines}
\usage{
makeLines(
  sPoly,
  xPoly = NULL,
  totalLengthKm,
  angle = 0,
  minLengthKm = 2,
  offset = 0,
  minSpace = NULL,
  maxSpace = NULL,
  optimTol = 0.01,
  outFile = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{sPoly}{\code{sf} object, the input polygon within which line transects
will be generated. The coordinate system
must use meters as the linear unit (e.g. EPSG=26913)}

\item{xPoly}{\code{sf} object, optional input polygon that defines an
exclusion area where line transects will not be generated.}

\item{totalLengthKm}{Total length of transects to generate (in kilometers)}

\item{angle}{Orientation of the transects. \code{angle = 0} produces
transects oriented North-South (the default), \code{angle = 90} produces
East-West transects.}

\item{minLengthKm}{Minimum length required for an individual transect line
(in kilometers)}

\item{offset}{An offset to use when placing transects. By default the central
transect is aligned with the center of the polygon.}

\item{minSpace}{Optional minimal line spacing (in kilometers) to consider when
optimizing transect placement (default is NULL).}

\item{maxSpace}{Optional maximum line spacing (in kilometers) to consider when
optimizing transect placement (default is NULL).}

\item{optimTol}{Optional tolerance value indicating the maximum allowable
deviation from \code{totalLengthKm} expressed as a proportion (default is 0.01).
When the function fails to find a solution within the specified tolerance an
error is returned.}

\item{outFile}{Optional output filename for a saving a spatial layer containing
the line transects. Typically, this will be a shapefile
(e.g. "C:/Temp/myLines.shp")
The file extension can be any accepted by \code{sf::st_write()}}

\item{overwrite}{If \code{outFile} is specified, and a file already exists
with that name, should that existing file be overwritten by the new one?
(default is TRUE)}
}
\value{
Returns a list object containing the transect lines and
a summary dataframe:
\describe{
  \item{lines}{\code{sf} object, simple feature collection of transect lines}
  \item{summary}{\code{data.frame} summary of transect lines}
}
}
\description{
This function optimizes the placement and spacing of
line transects for a given polygon area and desired total
transect length.
}
\examples{
\dontrun{
#Read in herd unit polygon
hu <- sf::st_read("W:/My Drive/Projects/Pronghorn_ltds/herd_units/Rattlesnake_HU.shp")

# Make lines
testLines <- makeLines(sPoly = hu,
     xPoly = NULL,
     totalLengthKm = 1000,
     angle = 0,
     outFile = NULL,
     minSpace = NULL,
     maxSpace = NULL)

# View output summary
print(testLines$summary)

# Plot output
plot(hu["geometry"])
plot(testLines$lines['lineID'], add = T, col = 'red')
}
}
\author{
Tom Prebyl, minor contributions by Jason Carlisle
}
