% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapLT.R
\name{mapLT}
\alias{mapLT}
\title{mapLT}
\usage{
mapLT(herdPoly, cutoutPoly = NULL, lines = NULL)
}
\arguments{
\item{herdPoly}{sf object, herd unit boundary (sample frame).}

\item{cutoutPoly}{sf object, area excluded from sample frame.}

\item{lines}{sf object, transect lines.}
}
\description{
quick wrapper for leaflet maps
}
