% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapResults.R
\name{mapResults}
\alias{mapResults}
\title{mapResults}
\usage{
mapResults(preppedData)
}
\arguments{
\item{preppedData}{List object returned from \code{prepDataForAnalysis}.}
}
\description{
quick wrapper for leaflet maps
}
