% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDataForAnalysis.R
\name{prepDataForAnalysis}
\alias{prepDataForAnalysis}
\title{prepDataForAnalysis}
\usage{
prepDataForAnalysis(inputFile = NULL, inputSheet = NULL, shpCRS = 26913)
}
\arguments{
\item{inputFile}{The survey data output from CyberTracker.  Excel files with
 extension \code{.xls} or \code{.xlsx} are
allowed. If \code{inputFile = NULL} (the default), the user is
prompted to select the input file interactively. Alternatively, the user can
specify the path and name of the Excel file containing the input data.
Note the direction of slashes in the path must follow R expectations
(use \code{/} instead of the Windows default of \code{\\}),
that the file name and extension must be included,
and that the path must start and end with quotes.
Example:  \code{"C:/Users/jcarlisle/Desktop/myInputFile.xlsx"}.
This function expects the following columns (named as shown) in the
\code{inputFile}.  The order of the columns does not matter (except in the
case of the "Cluster size" columns, see below), and
any additional columns are ignored:
\describe{
 \item{Transect number}{Unique name of each transect.}
 \item{Flight control}{Indicates transect starts/stops from detections.}
 \item{Latitude}{Y coordinate from GPS.}
 \item{Longitude}{X coordinate from GPS.}
 \item{Range}{Height of airplane (above ground level [AGL]) in feet.}
 \item{Cluster size}{For some reason, the CyberTracker output contains two
 columns named "Cluster size".  This function expects the first one to be
 the number of pronghorn in the detected group.}
 \item{Cluster size}{For some reason, the CyberTracker output contains two
 columns named "Cluster size".  This function expects the second one to be
 the distance band the pronghorn were detected in.}
 \item{Date}{Date of survey.  Can be blank, but issues a warning if so.}
 \item{Time}{Time of survey.  Can be blank, but issues a warning if so.}
}}

\item{inputSheet}{The sheet (or tab) within the above-specified Excel file
that should be read in. If \code{inputSheet = NULL} (the default), the names
of the sheets in the \code{inputFile} are printed to the console, and the
user is prompted to select the sheet name interactively.  Alternatively,
specify as the name of the sheet (e.g., \code{"Sheet1"} or \code{"inputData"})
or as the numbered position of the sheet within the file (e.g., \code{1}).}

\item{shpCRS}{Numeric, the EPSG code for the coordinate reference system
(projection) of the coordinates in the input data.  Common projections for
Wyoming:
\itemize{
\item 26912:  UTM NAD83 Zone 12N
\item 26913:  UTM NAD83 Zone 13N (default)
}}
}
\value{
Returns a list with six elements:
\describe{
 \item{ddf}{A data.frame of detection data.  One row for each group of
   pronghorn detected.  See the \code{detectionData} argument of
   \code{Rdistance::dfuncEstim} and \code{Rdistance::abundEstim}.}
 \item{sdf}{A data.frame of site data.  One row for each site (transect)
   surveyed.  See the \code{siteData} argument of
   \code{Rdistance::dfuncEstim} and \code{Rdistance::abundEstim}.}
 \item{txt}{A data.frame in the "flat file" format expected by Program
   DISTANCE.  Ready to be written to a tab-delimited text file which can be
   imported into Program DISTANCE.  The column names are
   standardized to use names similar to Program DISTANCE (making import less
   error-prone) and to clarify units:
   \describe{
   \item{lineLabel}{Unique name of each transect.}
   \item{lineLengthKm}{Length of each transect in kilometers.}
   \item{coordsX}{X coordinate from GPS.}
   \item{coordsY}{Y coordinate from GPS.}
   \item{flightHeightFt}{Height of airplane (AGL) in feet.}
   \item{clusterSize}{Number of pronghorn in detected group.}
   \item{distBand}{Distance band the pronghorn were detected in.}
   \item{adjustedDistM}{Estimated distance from transect to the detected
   pronghorn group in meters. Based on the midpoint of the distance band
   containing the detected group, adjusted for the flight height at the time
   of detection.}
   }}
 \item{sumTable}{A data.frame of summaries of the input dataset, meant to aid
   in QAQC and troubleshooting.}
 \item{sfLines}{An sf object of the transect lines as surveyed.
   The transects are reconstructed from the transect start and stop records
   in the input data.
   See the \code{shpCRS} argument to specify a coordinate reference system.}
 \item{sfPoints}{An sf object of the detection points.  The points are
   created from the sighting records in the input data.  Note, these
   locations represent the location of the plane when a pronghorn
   group was detected, not the location of the pronghorn group itself.
   See the \code{shpCRS} argument to specify a coordinate reference system.}
}
}
\description{
This function reads in raw survey data collected during
aerial line transect surveys for pronghorn (exported directly from
CyberTracker data-collection software) and prepares it for distance-sampling
analysis in Program R or Program DISTANCE.
}
\examples{
\dontrun{
# Read in Sheet1 of myInputFile.xlsx
x <- prepDataForAnalysis(
  inputFile = "C:/Users/myUserName/Desktop/myInputFile.xlsx",
  inputSheet = "Sheet1"
)

# Detection data (one row per group detected)
head(x$ddf)

# Site data (one row per transect surveyed)
head(x$sdf)

# Both detection and site data (flat file) ready for import to Program DISTANCE
head(x$txt)

# Summaries of the data (to aid QAQC)
x$sumTable

# Map of transects as flown
plot(sf::st_geometry(x$sfLines))

# Map of pronghorn detection locations (plane location)
plot(sf::st_geometry(x$sfPoints))

# Interactive map of both lines and points
# Points are sized based on group size
mapview::mapview(x$sfLines, color = "blue", legend = FALSE) +
  mapview::mapview(x$sfPoints, cex = "s", label = "s")
}
}
\references{
Guenzel, R.J. 2007. Procedures for Estimating Pronghorn Abundance
in Wyoming Using Aerial Line Transect Sampling. Wyoming Game and Fish Department.
Cheyenne, WY, USA.
}
\author{
Jason Carlisle
}
